fix: page performance
- remove parallel routes because too buggy
