+++ app/user/[id]/_user-top.tsx
@@ -1,10 +1,10 @@
 import { auth } from '@/app/api/auth/[...nextauth]/auth';
+import { MumbleUserCard } from '@/components/mumble/mumble-user-card';
 import { GET_USER_BY_ID } from '@/utils/api/api-actions-user';
 import { APIError } from '@/utils/api/api-service-base';
 import { notFound } from 'next/navigation';
-import { MumbleUserCard } from '@/components/mumble/mumble-user-card';
 
-export default async function Page({ params }: { params: { id: string } }) {
+export default async function UserTop({ params }: { params: { id: string } }) {
   const session = await auth();
   try {
     const user = await GET_USER_BY_ID({ id: params.id });
